requirements.txt
Django==1.5.4
Pillow==2.2.1
South==0.8.2
argparse==1.2.1
django-vanilla-views==1.0.2
psycopg2==2.5.1
wsgiref==0.1.2
